+++ lang/src/org/partiql/lang/eval/builtins/ExtractExprFunction.kt
@@ -25,13 +25,13 @@ import org.partiql.lang.eval.dateTimePartValue
 import org.partiql.lang.eval.dateValue
 import org.partiql.lang.eval.errNoContext
 import org.partiql.lang.eval.isUnknown
+import org.partiql.lang.eval.time.Time
 import org.partiql.lang.eval.timeValue
 import org.partiql.lang.eval.timestampValue
 import org.partiql.lang.syntax.DateTimePart
 import org.partiql.lang.types.AnyOfType
 import org.partiql.lang.types.FunctionSignature
 import org.partiql.lang.types.StaticType
-import org.partiql.lang.util.Time
 import java.math.BigDecimal
 import java.time.LocalDate
