+++ lang/src/org/partiql/lang/eval/builtins/MakeTimeExprFunction.kt
@@ -9,10 +9,10 @@ import org.partiql.lang.eval.ExprValueFactory
 import org.partiql.lang.eval.bigDecimalValue
 import org.partiql.lang.eval.err
 import org.partiql.lang.eval.intValue
+import org.partiql.lang.eval.time.NANOS_PER_SECOND
+import org.partiql.lang.eval.time.Time
 import org.partiql.lang.types.FunctionSignature
 import org.partiql.lang.types.StaticType
-import org.partiql.lang.util.NANOS_PER_SECOND
-import org.partiql.lang.util.Time
 import java.math.BigDecimal
 
 /**
