+++ lang/src/org/partiql/lang/eval/time/TimeExtensions.kt
@@ -1,6 +1,7 @@
-package org.partiql.lang.util
+package org.partiql.lang.eval.time
 
 import org.partiql.lang.errors.ErrorCode
+import org.partiql.lang.util.propertyValueMapOf
 import java.time.ZoneOffset
 import kotlin.math.absoluteValue
