+++ lang/src/org/partiql/lang/syntax/SqlParser.kt
@@ -37,17 +37,18 @@ import org.partiql.lang.domains.metaContainerOf
 import org.partiql.lang.errors.ErrorCode
 import org.partiql.lang.errors.Property
 import org.partiql.lang.errors.PropertyValueMap
+import org.partiql.lang.eval.time.DATE_PATTERN_REGEX
+import org.partiql.lang.eval.time.MAX_PRECISION_FOR_TIME
+import org.partiql.lang.eval.time.genericTimeRegex
+import org.partiql.lang.eval.time.getPrecisionFromTimeString
+import org.partiql.lang.eval.time.timeWithoutTimeZoneRegex
 import org.partiql.lang.types.CustomType
 import org.partiql.lang.util.BuiltInScalarTypeId
-import org.partiql.lang.util.DATE_PATTERN_REGEX
-import org.partiql.lang.util.MAX_PRECISION_FOR_TIME
 import org.partiql.lang.util.asIonInt
 import org.partiql.lang.util.atomFromHead
 import org.partiql.lang.util.checkThreadInterrupted
 import org.partiql.lang.util.err
 import org.partiql.lang.util.errExpectedTokenType
-import org.partiql.lang.util.genericTimeRegex
-import org.partiql.lang.util.getPrecisionFromTimeString
 import org.partiql.lang.util.head
 import org.partiql.lang.util.ionValue
 import org.partiql.lang.util.isNumeric
@@ -60,7 +61,6 @@ import org.partiql.lang.util.stringValue
 import org.partiql.lang.util.tail
 import org.partiql.lang.util.tailExpectedKeyword
 import org.partiql.lang.util.tailExpectedToken
-import org.partiql.lang.util.timeWithoutTimeZoneRegex
 import org.partiql.lang.util.unaryMinus
 import org.partiql.pig.runtime.LongPrimitive
 import org.partiql.pig.runtime.SymbolPrimitive
